Split AccountRow files
